fix file xóa bỏ

+++ Services/Repositories/Implimentations/QuanLyHoaDon/HoaDonDienTuService.cs
@@ -3132,12 +3132,10 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
                 if (hd.TrangThai == (int)TrangThaiHoaDon.HoaDonXoaBo)
                 {
                     string pathFilePDF = $"{_hostingEnvironment.WebRootPath}/FilesUpload/{databaseName}/{ManageFolderPath.PDF_SIGNED}/{hd.FileDaKy}";
-                    if (File.Exists(pathFilePDF))
-                    {
-                        File.Delete(pathFilePDF);
-                        var binPDF = await _db.FileDatas.Where(x => x.RefId == hd.HoaDonDienTuId && x.IsSigned == true).Select(x => x.Binary).FirstOrDefaultAsync();
-                        File.WriteAllBytes(pathFilePDF, binPDF);
-                    }
+                    if (File.Exists(pathFilePDF)) File.Delete(pathFilePDF);
+                    var binPDF = await _db.FileDatas.Where(x => x.RefId == hd.HoaDonDienTuId && x.IsSigned == true && x.Type == 2).Select(x => x.Binary).FirstOrDefaultAsync();
+                    File.WriteAllBytes(pathFilePDF, binPDF);
+                    await addTextDelete(pathFilePDF);
                 }
                 if (hd.IsCapMa != true && hd.IsReloadSignedPDF != true && hd.BuyerSigned != true && (hd.TrangThaiQuyTrinh >= (int)TrangThaiQuyTrinh.DaKyDienTu) && (hd.TrangThaiQuyTrinh != (int)TrangThaiQuyTrinh.GuiTCTNLoi) && (!string.IsNullOrEmpty(hd.FileDaKy) || !string.IsNullOrEmpty(hd.XMLDaKy)))
                 {
@@ -6946,7 +6944,30 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
                }).ToList();
             return enums;
         }
+        public async Task<bool> addTextDelete(string pdfFilePath)
+        {
+            //thêm ảnh đã bị xóa vào file pdf
+            if (@File.Exists(pdfFilePath))
+            {
+                string mauHoaDonImg = Path.Combine(_hostingEnvironment.WebRootPath, "images/template/dabixoabo.png");
+                PdfDocument pdfDoc = new PdfDocument();
+                pdfDoc.LoadFromFile(pdfFilePath);
+                PdfImage image = PdfImage.FromFile(mauHoaDonImg);
 
+                int pdfPageCount = pdfDoc.Pages.Count;
+                for (int i = 0; i < pdfPageCount; i++)
+                {
+                    PdfPageBase page = pdfDoc.Pages[i];
+                    page.Canvas.SetTransparency(0.7f, 0.7f, PdfBlendMode.Normal);
+                    page.Canvas.DrawImage(image, new PointF(130, 270), new SizeF(350, 350));
+                }
+
+                pdfDoc.SaveToFile(pdfFilePath);
+                pdfDoc.Close();
+
+            }
+            return true;
+        }
         [Obsolete]
         public async Task<bool> XoaBoHoaDon(ParamXoaBoHoaDon @params)
         {
@@ -6973,28 +6994,8 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
                         string pdfFilePath = Path.Combine(_hostingEnvironment.WebRootPath, pdfPath);
                         if (!@File.Exists(pdfFilePath))
                             await RestoreFilesInvoiceSigned(_objHDDT.HoaDonDienTuId);
-                        //thêm ảnh đã bị xóa vào file pdf
-                        if (@File.Exists(pdfFilePath))
-                        {
-                            string mauHoaDonImg = Path.Combine(_hostingEnvironment.WebRootPath, "images/template/dabixoabo.png");
-                            PdfDocument pdfDoc = new PdfDocument();
-                            pdfDoc.LoadFromFile(pdfFilePath);
-                            PdfImage image = PdfImage.FromFile(mauHoaDonImg);
 
-                            int pdfPageCount = pdfDoc.Pages.Count;
-                            for (int i = 0; i < pdfPageCount; i++)
-                            {
-                                PdfPageBase page = pdfDoc.Pages[i];
-                                page.Canvas.SetTransparency(0.7f, 0.7f, PdfBlendMode.Normal);
-                                page.Canvas.DrawImage(image, new PointF(130, 270), new SizeF(350, 350));
-                            }
-
-                            pdfDoc.SaveToFile(pdfFilePath);
-                            pdfDoc.Close();
-
-                            //Update file pdf fileData
-                            await UpdateFileDataPdfForHDDT(_objHDDT.HoaDonDienTuId, pdfFilePath);
-                        }
+                        await addTextDelete(pdfFilePath);
                     }
                     if (@params.OptionalSend == 1)
                     {
